+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+* Reconnect to inverter after 15 minutes of not receiving any data (#223)
+* Fix max/min cell temperature/voltage decoding as reported from BMS (#227)
+* Add more HA entities: max/min cell temp/voltage, more charge powers (#228)
+* Add ReadInput4 with EG4 18k generator data (#239, @pmccut)
+
+
 # 0.13.0 - 27th October 2023
 
 * **BREAKING CHANGE**: Simplify scheduler.timesync configuration to appease HA (#209)